+++ source/isaaclab/isaaclab/envs/mdp/events.py
@@ -17,14 +17,12 @@ from __future__ import annotations
 import logging
 import math
 import re
-from typing import TYPE_CHECKING, Literal
+from typing import TYPE_CHECKING
 
 import torch
-import warp as wp
 
 import isaaclab.sim as sim_utils
 import isaaclab.utils.math as math_utils
-from isaaclab.actuators import ImplicitActuator
 from isaaclab.managers import EventTermCfg, ManagerTermBase, SceneEntityCfg
 from isaaclab.utils.version import compare_versions
 
@@ -37,1665 +35,6 @@ if TYPE_CHECKING:
 
 # import logger
 logger = logging.getLogger(__name__)
-
-
-def randomize_rigid_body_scale(
-    env: ManagerBasedEnv,
-    env_ids: torch.Tensor | None,
-    scale_range: tuple[float, float] | dict[str, tuple[float, float]],
-    asset_cfg: SceneEntityCfg,
-    relative_child_path: str | None = None,
-):
-    """Randomize the scale of a rigid body asset in the USD stage.
-
-    This function modifies the "xformOp:scale" property of all the prims corresponding to the asset.
-
-    It takes a tuple or dictionary for the scale ranges. If it is a tuple, then the scaling along
-    individual axis is performed equally. If it is a dictionary, the scaling is independent across each dimension.
-    The keys of the dictionary are ``x``, ``y``, and ``z``. The values are tuples of the form ``(min, max)``.
-
-    If the dictionary does not contain a key, the range is set to one for that axis.
-
-    Relative child path can be used to randomize the scale of a specific child prim of the asset.
-    For example, if the asset at prim path expression ``/World/envs/env_.*/Object`` has a child
-    with the path ``/World/envs/env_.*/Object/mesh``, then the relative child path should be ``mesh`` or
-    ``/mesh``.
-
-    .. attention::
-        Since this function modifies USD properties that are parsed by the physics engine once the simulation
-        starts, the term should only be used before the simulation starts playing. This corresponds to the
-        event mode named "usd". Using it at simulation time, may lead to unpredictable behaviors.
-
-    .. note::
-        When randomizing the scale of individual assets, please make sure to set
-        :attr:`isaaclab.scene.InteractiveSceneCfg.replicate_physics` to False. This ensures that physics
-        parser will parse the individual asset properties separately.
-    """
-    # check if sim is running
-    if env.sim.is_playing():
-        raise RuntimeError(
-            "Randomizing scale while simulation is running leads to unpredictable behaviors."
-            " Please ensure that the event term is called before the simulation starts by using the 'usd' mode."
-        )
-
-    # extract the used quantities (to enable type-hinting)
-    asset: RigidObject = env.scene[asset_cfg.name]
-
-    if any(cls.__name__ == "Articulation" for cls in type(asset).__mro__):
-        raise ValueError(
-            "Scaling an articulation randomly is not supported, as it affects joint attributes and can cause"
-            " unexpected behavior. To achieve different scales, we recommend generating separate USD files for"
-            " each version of the articulation and using multi-asset spawning. For more details, refer to:"
-            " https://isaac-sim.github.io/IsaacLab/main/source/how-to/multi_asset_spawning.html"
-        )
-
-    # resolve environment ids
-    if env_ids is None:
-        env_ids = torch.arange(env.scene.num_envs, device="cpu")
-    else:
-        env_ids = env_ids.cpu()
-
-    # acquire stage
-    stage = env.sim.stage
-    # resolve prim paths for spawning and cloning
-    prim_paths = sim_utils.find_matching_prim_paths(asset.cfg.prim_path)
-
-    # sample scale values
-    if isinstance(scale_range, dict):
-        range_list = [scale_range.get(key, (1.0, 1.0)) for key in ["x", "y", "z"]]
-        ranges = torch.tensor(range_list, device="cpu")
-        rand_samples = math_utils.sample_uniform(ranges[:, 0], ranges[:, 1], (len(env_ids), 3), device="cpu")
-    else:
-        rand_samples = math_utils.sample_uniform(*scale_range, (len(env_ids), 1), device="cpu")
-        rand_samples = rand_samples.repeat(1, 3)
-    # convert to list for the for loop
-    rand_samples = rand_samples.tolist()
-
-    # apply the randomization to the parent if no relative child path is provided
-    # this might be useful if user wants to randomize a particular mesh in the prim hierarchy
-    if relative_child_path is None:
-        relative_child_path = ""
-    elif not relative_child_path.startswith("/"):
-        relative_child_path = "/" + relative_child_path
-
-    # use sdf changeblock for faster processing of USD properties (local: pxr only available with Kit)
-    from pxr import Gf, Sdf, UsdGeom, Vt  # noqa: PLC0415
-
-    with Sdf.ChangeBlock():
-        for i, env_id in enumerate(env_ids):
-            # path to prim to randomize
-            prim_path = prim_paths[env_id] + relative_child_path
-            # spawn single instance
-            prim_spec = Sdf.CreatePrimInLayer(stage.GetRootLayer(), prim_path)
-
-            # get the attribute to randomize
-            scale_spec = prim_spec.GetAttributeAtPath(prim_path + ".xformOp:scale")
-            # if the scale attribute does not exist, create it
-            has_scale_attr = scale_spec is not None
-            if not has_scale_attr:
-                scale_spec = Sdf.AttributeSpec(prim_spec, prim_path + ".xformOp:scale", Sdf.ValueTypeNames.Double3)
-
-            # set the new scale
-            scale_spec.default = Gf.Vec3f(*rand_samples[i])
-
-            # ensure the operation is done in the right ordering if we created the scale attribute.
-            # otherwise, we assume the scale attribute is already in the right order.
-            # note: by default isaac sim follows this ordering for the transform stack so any asset
-            #   created through it will have the correct ordering
-            if not has_scale_attr:
-                op_order_spec = prim_spec.GetAttributeAtPath(prim_path + ".xformOpOrder")
-                if op_order_spec is None:
-                    op_order_spec = Sdf.AttributeSpec(
-                        prim_spec, UsdGeom.Tokens.xformOpOrder, Sdf.ValueTypeNames.TokenArray
-                    )
-                op_order_spec.default = Vt.TokenArray(["xformOp:translate", "xformOp:orient", "xformOp:scale"])
-
-
-class _RandomizeRigidBodyMaterialPhysx:
-    """PhysX backend implementation for material randomization.
-
-    Uses the bucket-based approach required by PhysX's 64000 unique material limit.
-    Materials are pre-sampled into buckets and randomly assigned to shapes.
-    """
-
-    def __init__(
-        self, cfg: EventTermCfg, env: ManagerBasedEnv, asset: RigidObject | Articulation, asset_cfg: SceneEntityCfg
-    ):
-        from isaaclab.assets import BaseArticulation
-
-        # obtain parameters for sampling friction and restitution values
-        static_friction_range = cfg.params.get("static_friction_range", (1.0, 1.0))
-        dynamic_friction_range = cfg.params.get("dynamic_friction_range", (1.0, 1.0))
-        restitution_range = cfg.params.get("restitution_range", (0.0, 0.0))
-        num_buckets = int(cfg.params.get("num_buckets", 1))
-
-        # sample material properties from the given ranges
-        # note: we only sample the materials once during initialization
-        #   afterwards these are randomly assigned to the geometries of the asset
-        range_list = [static_friction_range, dynamic_friction_range, restitution_range]
-        ranges = torch.tensor(range_list, device="cpu")
-        self.material_buckets = math_utils.sample_uniform(ranges[:, 0], ranges[:, 1], (num_buckets, 3), device="cpu")
-
-        # ensure dynamic friction is always less than static friction
-        make_consistent = cfg.params.get("make_consistent", False)
-        if make_consistent:
-            self.material_buckets[:, 1] = torch.min(self.material_buckets[:, 0], self.material_buckets[:, 1])
-
-        self.asset = asset
-        self.asset_cfg = asset_cfg
-
-        # obtain number of shapes per body (needed for indexing the material properties correctly)
-        # note: this is a workaround since the Articulation does not provide a direct way to obtain the number of shapes
-        #  per body. We use the physics simulation view to obtain the number of shapes per body.
-        if isinstance(asset, BaseArticulation) and asset_cfg.body_ids != slice(None):
-            self.num_shapes_per_body = []
-            for link_path in asset.root_view.link_paths[0]:
-                link_physx_view = asset._physics_sim_view.create_rigid_body_view(link_path)  # type: ignore
-                self.num_shapes_per_body.append(link_physx_view.max_shapes)
-            # ensure the parsing is correct
-            num_shapes = sum(self.num_shapes_per_body)
-            expected_shapes = asset.root_view.max_shapes
-            if num_shapes != expected_shapes:
-                raise ValueError(
-                    "Randomization term 'randomize_rigid_body_material' failed to parse the number of shapes per body."
-                    f" Expected total shapes: {expected_shapes}, but got: {num_shapes}."
-                )
-        else:
-            # in this case, we don't need to do special indexing
-            self.num_shapes_per_body = None
-
-    def __call__(
-        self,
-        env: ManagerBasedEnv,
-        env_ids: torch.Tensor | None,
-        static_friction_range: tuple[float, float],
-        dynamic_friction_range: tuple[float, float],
-        restitution_range: tuple[float, float],
-        num_buckets: int,
-        asset_cfg: SceneEntityCfg,
-        make_consistent: bool = False,
-    ):
-        # resolve environment ids
-        if env_ids is None:
-            env_ids = torch.arange(env.scene.num_envs, device="cpu", dtype=torch.int32)
-        else:
-            env_ids = env_ids.cpu()
-
-        # randomly assign material IDs to the geometries
-        total_num_shapes = self.asset.root_view.max_shapes
-        bucket_ids = torch.randint(0, num_buckets, (len(env_ids), total_num_shapes), device="cpu")
-        material_samples = self.material_buckets[bucket_ids]
-
-        # retrieve material buffer from the physics simulation
-        materials = wp.to_torch(self.asset.root_view.get_material_properties())
-
-        # update material buffer with new samples
-        if self.num_shapes_per_body is not None:
-            # sample material properties from the given ranges
-            for body_id in self.asset_cfg.body_ids:
-                # obtain indices of shapes for the body
-                start_idx = sum(self.num_shapes_per_body[:body_id])
-                end_idx = start_idx + self.num_shapes_per_body[body_id]
-                # assign the new materials
-                # material samples are of shape: num_env_ids x total_num_shapes x 3
-                materials[env_ids, start_idx:end_idx] = material_samples[:, start_idx:end_idx]
-        else:
-            # assign all the materials
-            materials[env_ids] = material_samples[:]
-
-        # apply to simulation
-        self.asset.root_view.set_material_properties(
-            wp.from_torch(materials, dtype=wp.float32), wp.from_torch(env_ids, dtype=wp.int32)
-        )
-
-
-class _RandomizeRigidBodyMaterialNewton:
-    """Newton backend implementation for material randomization.
-
-    Newton can assign arbitrary friction/restitution per shape (no bucket limitation).
-    Samples friction (mu) and restitution continuously from the given ranges.
-    Newton uses a single friction coefficient (mu), so ``dynamic_friction_range``
-    and ``num_buckets`` are ignored.
-    """
-
-    def __init__(
-        self, cfg: EventTermCfg, env: ManagerBasedEnv, asset: RigidObject | Articulation, asset_cfg: SceneEntityCfg
-    ):
-        import isaaclab_newton.physics.newton_manager as newton_manager_module  # noqa: PLC0415
-        from isaaclab_newton.assets import Articulation as NewtonArticulation  # noqa: PLC0415
-        from newton.solvers import SolverNotifyFlags  # noqa: PLC0415
-
-        self.asset = asset
-        self.asset_cfg = asset_cfg
-        self._newton_manager = newton_manager_module.NewtonManager
-        self._notify_shape_properties = SolverNotifyFlags.SHAPE_PROPERTIES
-
-        # cache friction/restitution ranges for continuous per-shape sampling
-        self._static_friction_range = cfg.params.get("static_friction_range", (1.0, 1.0))
-        self._restitution_range = cfg.params.get("restitution_range", (0.0, 0.0))
-
-        # get friction/restitution view-level bindings
-        model = self._newton_manager.get_model()
-        self._friction_binding = asset._root_view.get_attribute("shape_material_mu", model)[:, 0]  # type: ignore
-        self._restitution_binding = asset._root_view.get_attribute("shape_material_restitution", model)[:, 0]  # type: ignore
-
-        # compute shape indices for body-specific randomization
-        if isinstance(asset, NewtonArticulation) and asset_cfg.body_ids != slice(None):
-            num_shapes_per_body = asset.num_shapes_per_body
-            shape_indices_list = []
-            for body_id in asset_cfg.body_ids:
-                start_idx = sum(num_shapes_per_body[:body_id])
-                end_idx = start_idx + num_shapes_per_body[body_id]
-                shape_indices_list.extend(range(start_idx, end_idx))
-            self._shape_indices = torch.tensor(shape_indices_list, dtype=torch.long)
-        else:
-            self._shape_indices = torch.arange(self._friction_binding.shape[1], dtype=torch.long)
-
-    def __call__(
-        self,
-        env: ManagerBasedEnv,
-        env_ids: torch.Tensor | None,
-        static_friction_range: tuple[float, float],
-        dynamic_friction_range: tuple[float, float],
-        restitution_range: tuple[float, float],
-        num_buckets: int,
-        asset_cfg: SceneEntityCfg,
-        make_consistent: bool = False,
-    ):
-        device = env.device
-        # resolve environment ids
-        if env_ids is None:
-            env_ids = torch.arange(env.scene.num_envs, device=device, dtype=torch.int32)
-        else:
-            env_ids = env_ids.to(device)
-
-        num_shapes = len(self._shape_indices)
-        shape_idx = self._shape_indices.to(device)
-
-        # sample friction (mu) and restitution continuously per shape
-        friction_range = torch.tensor(self._static_friction_range, device=device)
-        restitution_range_t = torch.tensor(self._restitution_range, device=device)
-        friction_samples = math_utils.sample_uniform(
-            friction_range[0], friction_range[1], (len(env_ids), num_shapes), device=device
-        )
-        restitution_samples = math_utils.sample_uniform(
-            restitution_range_t[0], restitution_range_t[1], (len(env_ids), num_shapes), device=device
-        )
-
-        # write only the affected env_ids to the warp binding
-        friction_view = wp.to_torch(self._friction_binding)
-        restitution_view = wp.to_torch(self._restitution_binding)
-        friction_view[env_ids[:, None], shape_idx] = friction_samples
-        restitution_view[env_ids[:, None], shape_idx] = restitution_samples
-
-        # notify the physics engine
-        self._newton_manager.add_model_change(self._notify_shape_properties)
-
-
-class randomize_rigid_body_material(ManagerTermBase):
-    """Randomize the physics materials on all geometries of the asset.
-
-    This function creates a set of physics materials with random static friction, dynamic friction, and restitution
-    values and assigns them to the geometries of the asset.
-
-    Automatically detects the active physics backend (PhysX or Newton) and delegates to
-    the appropriate backend-specific implementation:
-
-    - **PhysX**: Uses the 3-tuple material format (static_friction, dynamic_friction, restitution)
-      with bucket-based assignment (limited to 64000 unique materials). Applied via the PhysX
-      tensor API (``root_view.set_material_properties``).
-    - **Newton**: Samples friction (mu) and restitution continuously per shape (no bucket
-      limitation). Newton uses a single friction coefficient, so ``dynamic_friction_range``
-      and ``num_buckets`` are ignored. Applied directly to Newton's view-level bindings.
-
-    If the flag ``make_consistent`` is set to ``True``, the dynamic friction is set to be less than or equal to
-    the static friction (PhysX only). This obeys the physics constraint on friction values.
-
-    .. attention::
-        On PhysX, this function uses CPU tensors to assign the material properties. It is recommended to
-        use this function only during the initialization of the environment.
-
-    .. note::
-        PhysX only allows 64000 unique physics materials in the scene. If the number of materials exceeds this
-        limit, the simulation will crash. Due to this reason, we sample the materials only once during initialization.
-        Afterwards, these materials are randomly assigned to the geometries of the asset.
-    """
-
-    def __init__(self, cfg: EventTermCfg, env: ManagerBasedEnv):
-        """Initialize the term.
-
-        Args:
-            cfg: The configuration of the event term.
-            env: The environment instance.
-
-        Raises:
-            ValueError: If the asset is not a RigidObject or an Articulation.
-        """
-        from isaaclab.assets import BaseArticulation, BaseRigidObject
-
-        super().__init__(cfg, env)
-
-        # extract the used quantities (to enable type-hinting)
-        self.asset_cfg: SceneEntityCfg = cfg.params["asset_cfg"]
-        self.asset: RigidObject | Articulation = env.scene[self.asset_cfg.name]
-
-        if not isinstance(self.asset, (BaseRigidObject, BaseArticulation)):
-            raise ValueError(
-                f"Randomization term 'randomize_rigid_body_material' not supported for asset: '{self.asset_cfg.name}'"
-                f" with type: '{type(self.asset)}'."
-            )
-
-        # detect physics backend and instantiate the appropriate implementation.
-        # Check ``ovphysxmanager`` first: it contains the substring ``physx`` so
-        # would otherwise be caught by the ``"physx" in ...`` branch below and
-        # routed to the PhysX impl, which assumes a ``root_view`` with
-        # ``.link_paths`` — OVPhysX's per-tensor-type bindings dict does not
-        # satisfy that contract.  Newton's subclasses (``NewtonMJWarpManager``,
-        # ``NewtonKaminoManager``, ...) are caught by the substring branch.
-        manager_name = env.sim.physics_manager.__name__.lower()
-        if manager_name == "ovphysxmanager":
-            # No OVPhysX implementation yet — wheel-side
-            # ``RIGID_BODY_MATERIAL`` tensor binding is missing; randomization
-            # would require per-body view creation that ovphysx does not yet
-            # expose.  Run with material randomization disabled (warns once).
-            import logging  # noqa: PLC0415
-
-            logging.getLogger(__name__).warning(
-                "randomize_rigid_body_material is a no-op on the OVPhysX backend "
-                "(wheel-side gap — see docs/superpowers/specs/2026-04-27-ovphysx-contact-api-gaps.md)."
-            )
-
-            class _Noop:
-                def __call__(self, *args, **kwargs):
-                    pass
-
-            self._impl = _Noop()
-        elif "newton" in manager_name:
-            self._impl = _RandomizeRigidBodyMaterialNewton(cfg, env, self.asset, self.asset_cfg)
-        elif "physx" in manager_name:
-            self._impl = _RandomizeRigidBodyMaterialPhysx(cfg, env, self.asset, self.asset_cfg)
-        else:
-            raise ValueError(f"Unsupported physics manager for randomize_rigid_body_material: {manager_name!r}")
-
-    def __call__(
-        self,
-        env: ManagerBasedEnv,
-        env_ids: torch.Tensor | None,
-        static_friction_range: tuple[float, float],
-        dynamic_friction_range: tuple[float, float],
-        restitution_range: tuple[float, float],
-        num_buckets: int,
-        asset_cfg: SceneEntityCfg,
-        make_consistent: bool = False,
-    ):
-        self._impl(
-            env,
-            env_ids,
-            static_friction_range,
-            dynamic_friction_range,
-            restitution_range,
-            num_buckets,
-            asset_cfg,
-            make_consistent,
-        )
-
-
-class randomize_rigid_body_mass(ManagerTermBase):
-    """Randomize the mass of the bodies by adding, scaling, or setting random values.
-
-    This function allows randomizing the mass of the bodies of the asset. The function samples random
-    values from the given distribution parameters and adds, scales, or sets the values into the physics
-    simulation based on the operation.
-
-    If the :attr:`recompute_inertia` flag is set to :obj:`True`, the function recomputes the inertia tensor
-    of the bodies after setting the mass. This is useful when the mass is changed significantly, as the
-    inertia tensor depends on the mass. It assumes the body is a uniform density object. If the body is not
-    a uniform density object, the inertia tensor may not be accurate.
-
-    .. tip::
-        This function uses CPU tensors to assign the body masses. It is recommended to use this function
-        only during the initialization of the environment.
-    """
-
-    def __init__(self, cfg: EventTermCfg, env: ManagerBasedEnv):
-        """Initialize the term.
-
-        Args:
-            cfg: The configuration of the event term.
-            env: The environment instance.
-
-        Raises:
-            TypeError: If `params` is not a tuple of two numbers.
-            ValueError: If the operation is not supported.
-            ValueError: If the lower bound is negative or zero when not allowed.
-            ValueError: If the upper bound is less than the lower bound.
-        """
-        super().__init__(cfg, env)
-
-        # extract the used quantities (to enable type-hinting)
-        self.asset_cfg: SceneEntityCfg = cfg.params["asset_cfg"]
-        self.asset: RigidObject | Articulation = env.scene[self.asset_cfg.name]
-        # check for valid operation
-        if cfg.params["operation"] == "scale":
-            if "mass_distribution_params" in cfg.params:
-                _validate_scale_range(
-                    cfg.params["mass_distribution_params"], "mass_distribution_params", allow_zero=False
-                )
-        elif cfg.params["operation"] not in ("abs", "add"):
-            raise ValueError(
-                "Randomization term 'randomize_rigid_body_mass' does not support operation:"
-                f" '{cfg.params['operation']}'."
-            )
-        if cfg.params.get("min_mass") is not None:
-            if cfg.params.get("min_mass") < 1e-6:
-                raise ValueError(
-                    "Randomization term 'randomize_rigid_body_mass' does not support 'min_mass' less than 1e-6 to avoid"
-                    " physics errors."
-                )
-
-        self.default_mass = None
-        self.default_inertia = None
-
-    def __call__(
-        self,
-        env: ManagerBasedEnv,
-        env_ids: torch.Tensor | None,
-        asset_cfg: SceneEntityCfg,
-        mass_distribution_params: tuple[float, float],
-        operation: Literal["add", "scale", "abs"],
-        distribution: Literal["uniform", "log_uniform", "gaussian"] = "uniform",
-        recompute_inertia: bool = True,
-        min_mass: float = 1e-6,
-    ):
-        if self.default_mass is None:
-            self.default_mass = self.asset.data.body_mass.torch.clone()
-        if self.default_inertia is None:
-            self.default_inertia = self.asset.data.body_inertia.torch.clone()
-        # resolve environment ids
-        if env_ids is None:
-            env_ids = torch.arange(env.scene.num_envs, device=self.asset.device, dtype=torch.int32)
-        else:
-            env_ids = env_ids.to(self.asset.device)
-
-        # resolve body indices
-        if self.asset_cfg.body_ids == slice(None):
-            body_ids = torch.arange(self.asset.num_bodies, dtype=torch.int32, device=self.asset.device)
-        else:
-            body_ids = torch.tensor(self.asset_cfg.body_ids, dtype=torch.int32, device=self.asset.device)
-
-        # get the current masses of the bodies (num_assets, num_bodies)
-        masses = self.asset.data.body_mass.torch.clone()
-
-        # apply randomization on default values
-        # this is to make sure when calling the function multiple times, the randomization is applied on the
-        # default values and not the previously randomized values
-        masses[env_ids[:, None], body_ids] = self.default_mass[env_ids[:, None], body_ids].clone()
-
-        # sample from the given range
-        # note: we modify the masses in-place for all environments
-        #   however, the setter takes care that only the masses of the specified environments are modified
-        masses = _randomize_prop_by_op(
-            masses, mass_distribution_params, env_ids, body_ids, operation=operation, distribution=distribution
-        )
-        masses = torch.clamp(masses, min=min_mass)  # ensure masses are positive
-
-        # set the mass into the physics simulation
-        # note: backends expect partial data of shape (len(env_ids), len(body_ids))
-        self.asset.set_masses_index(masses=masses[env_ids[:, None], body_ids], body_ids=body_ids, env_ids=env_ids)
-
-        # recompute inertia tensors if needed
-        if recompute_inertia:
-            # compute the ratios of the new masses to the initial masses
-            ratios = masses[env_ids[:, None], body_ids] / self.default_mass[env_ids[:, None], body_ids]
-            # scale the inertia tensors by the the ratios
-            # since mass randomization is done on default values, we can use the default inertia tensors
-            inertias = self.asset.data.body_inertia.torch.clone()
-            # inertia has shape: (num_envs, num_bodies, 9) for all assets
-            inertias[env_ids[:, None], body_ids] = self.default_inertia[env_ids[:, None], body_ids] * ratios[..., None]
-            # set the inertia tensors into the physics simulation
-            # note: backends expect partial data of shape (len(env_ids), len(body_ids), 9)
-            self.asset.set_inertias_index(
-                inertias=inertias[env_ids[:, None], body_ids], body_ids=body_ids, env_ids=env_ids
-            )
-
-
-class randomize_rigid_body_inertia(ManagerTermBase):
-    """Randomize the inertia tensor of rigid bodies by adding, scaling, or setting values.
-
-    This function modifies body inertia tensors independently of mass. The inertia tensor
-    is a 3x3 symmetric matrix stored as 9 elements: ``[Ixx, Ixy, Ixz, Iyx, Iyy, Iyz, Izx, Izy, Izz]``.
-
-    Two modes are supported via the :attr:`diagonal_only` parameter:
-
-    - **diagonal_only=True** (default): Only modifies diagonal elements (Ixx, Iyy, Izz at
-      indices 0, 4, 8). This is useful for adding numerical stability (armature/regularization)
-      without changing rotational coupling between axes. The diagonal elements represent
-      resistance to rotation about each principal axis.
-
-    - **diagonal_only=False**: Modifies all 9 elements of the inertia tensor. This can
-      simulate manufacturing variations or asymmetric mass distributions. Off-diagonal
-      elements represent coupling between rotations about different axes.
-
-    .. note::
-        Unlike :class:`randomize_rigid_body_mass` which recomputes inertia based on mass
-        ratios, this function modifies inertia directly without affecting mass.
-    """
-
-    def __init__(self, cfg: EventTermCfg, env: ManagerBasedEnv):
-        """Initialize the term.
-
-        Args:
-            cfg: The configuration of the event term.
-            env: The environment instance.
-
-        Raises:
-            ValueError: If the operation is not supported.
-            ValueError: If the lower bound is negative or zero when not allowed for scale operation.
-            ValueError: If the upper bound is less than the lower bound.
-        """
-        from isaaclab.assets import BaseArticulation, BaseRigidObject, BaseRigidObjectCollection
-
-        super().__init__(cfg, env)
-
-        # extract the used quantities (to enable type-hinting)
-        self.asset_cfg: SceneEntityCfg = cfg.params["asset_cfg"]
-        self.asset: RigidObject | Articulation = env.scene[self.asset_cfg.name]
-
-        if not isinstance(self.asset, (BaseRigidObject, BaseArticulation, BaseRigidObjectCollection)):
-            raise ValueError(
-                f"Randomization term 'randomize_rigid_body_inertia' not supported for asset: '{self.asset_cfg.name}'"
-                f" with type: '{type(self.asset)}'."
-            )
-
-        # check for valid operation
-        if cfg.params["operation"] == "scale":
-            if "inertia_distribution_params" in cfg.params:
-                _validate_scale_range(
-                    cfg.params["inertia_distribution_params"], "inertia_distribution_params", allow_zero=False
-                )
-        elif cfg.params["operation"] not in ("abs", "add"):
-            raise ValueError(
-                f"Randomization term 'randomize_rigid_body_inertia' does not support operation:"
-                f" '{cfg.params['operation']}'."
-            )
-
-        self.default_inertia = None
-        # cache inertia indices: diagonal (0, 4, 8) for regularization, or all elements
-        diagonal_only = cfg.params.get("diagonal_only", True)
-        self._inertia_idx = torch.tensor([0, 4, 8], device=self.asset.device) if diagonal_only else slice(None)
-
-    def __call__(
-        self,
-        env: ManagerBasedEnv,
-        env_ids: torch.Tensor | None,
-        asset_cfg: SceneEntityCfg,
-        inertia_distribution_params: tuple[float, float],
-        operation: Literal["add", "scale", "abs"] = "add",
-        distribution: Literal["uniform", "log_uniform", "gaussian"] = "uniform",
-        diagonal_only: bool = True,
-    ):
-        """Randomize body inertia tensors.
-
-        Args:
-            env: The environment instance.
-            env_ids: The environment indices to randomize. If None, all environments are randomized.
-            asset_cfg: The asset configuration specifying the asset and body names.
-            inertia_distribution_params: Distribution parameters as a tuple of two floats
-                ``(low, high)`` for sampling inertia modification values.
-            operation: The operation to apply. Options: ``"add"``, ``"scale"``, ``"abs"``.
-                Defaults to ``"add"`` which is typical for regularization/armature.
-            distribution: The distribution to sample from. Options: ``"uniform"``,
-                ``"log_uniform"``, ``"gaussian"``. Defaults to ``"uniform"``.
-            diagonal_only: If True, only modify diagonal elements (Ixx, Iyy, Izz) for
-                numerical stability. If False, modify all 9 elements. Defaults to True.
-        """
-        # store default inertia on first call for repeatable randomization
-        if self.default_inertia is None:
-            self.default_inertia = self.asset.data.body_inertia.torch.clone()
-
-        # resolve environment ids
-        if env_ids is None:
-            env_ids = torch.arange(env.scene.num_envs, device=self.asset.device, dtype=torch.int32)
-        else:
-            env_ids = env_ids.to(self.asset.device)
-
-        # resolve body indices
-        if self.asset_cfg.body_ids == slice(None):
-            body_ids = torch.arange(self.asset.num_bodies, dtype=torch.int32, device=self.asset.device)
-        else:
-            body_ids = torch.tensor(self.asset_cfg.body_ids, dtype=torch.int32, device=self.asset.device)
-
-        # get default inertias for affected envs/bodies (advanced indexing creates a copy)
-        # shape: (len(env_ids), len(body_ids), 9)
-        inertias = self.default_inertia[env_ids[:, None], body_ids]
-
-        # resolve the distribution function
-        if distribution == "uniform":
-            dist_fn = math_utils.sample_uniform
-        elif distribution == "log_uniform":
-            dist_fn = math_utils.sample_log_uniform
-        elif distribution == "gaussian":
-            dist_fn = math_utils.sample_gaussian
-        else:
-            raise NotImplementedError(
-                f"Unknown distribution: '{distribution}' for inertia randomization."
-                " Please use 'uniform', 'log_uniform', or 'gaussian'."
-            )
-
-        # sample random values once per (env, body) - shape: (len(env_ids), len(body_ids))
-        random_values = dist_fn(*inertia_distribution_params, (len(env_ids), len(body_ids)), device=self.asset.device)
-
-        # apply the operation with the SAME random value per body
-        if operation == "add":
-            inertias[:, :, self._inertia_idx] += random_values[..., None]
-        elif operation == "scale":
-            inertias[:, :, self._inertia_idx] *= random_values[..., None]
-        elif operation == "abs":
-            inertias[:, :, self._inertia_idx] = random_values[..., None]
-
-        # set the inertia tensors into the physics simulation
-        self.asset.set_inertias_index(inertias=inertias, body_ids=body_ids, env_ids=env_ids)
-
-
-class randomize_rigid_body_com(ManagerTermBase):
-    """Randomize the center of mass (CoM) of rigid bodies by adding a random value sampled from the given ranges.
-
-    This class tracks the original CoM values and randomizes from those defaults on each call,
-    ensuring repeatable randomization across resets.
-
-    Automatically detects the active physics backend:
-
-    - **PhysX**: Passes the full CoM pose (position + quaternion) to ``set_coms_index``.
-    - **Newton**: Passes position-only (vec3) to ``set_coms_index``. Note that on Newton
-      (MuJoCo Warp), runtime CoM changes may cause simulation instability because
-      ``notify_model_changed(BODY_INERTIAL_PROPERTIES)`` does not fully recompute the
-      mass matrix after ``body_ipos`` changes. Use with caution until this is fixed upstream.
-    """
-
-    def __init__(self, cfg: EventTermCfg, env: ManagerBasedEnv):
-        """Initialize the term.
-
-        Args:
-            cfg: The configuration of the event term.
-            env: The environment instance.
-        """
-        super().__init__(cfg, env)
-
-        self.asset_cfg: SceneEntityCfg = cfg.params["asset_cfg"]
-        self.asset: RigidObject | Articulation = env.scene[self.asset_cfg.name]
-
-        # detect physics backend
-        manager_name = env.sim.physics_manager.__name__.lower()
-        self._is_newton = "newton" in manager_name
-
-        self.default_com = None
-
-    def __call__(
-        self,
-        env: ManagerBasedEnv,
-        env_ids: torch.Tensor | None,
-        com_range: dict[str, tuple[float, float]],
-        asset_cfg: SceneEntityCfg,
-    ):
-        # store default CoM on first call for repeatable randomization
-        if self.default_com is None:
-            self.default_com = self.asset.data.body_com_pose_b.torch.clone()
-
-        # resolve environment ids
-        if env_ids is None:
-            env_ids = torch.arange(env.scene.num_envs, device=self.asset.device)
-        else:
-            env_ids = env_ids.to(self.asset.device)
-
-        # resolve body indices
-        if self.asset_cfg.body_ids == slice(None):
-            body_ids = torch.arange(self.asset.num_bodies, dtype=torch.int, device=self.asset.device)
-        else:
-            body_ids = torch.tensor(self.asset_cfg.body_ids, dtype=torch.int, device=self.asset.device)
-
-        # sample random CoM values
-        range_list = [com_range.get(key, (0.0, 0.0)) for key in ["x", "y", "z"]]
-        ranges = torch.tensor(range_list, device=self.asset.device)
-        rand_samples = math_utils.sample_uniform(
-            ranges[:, 0], ranges[:, 1], (len(env_ids), 3), device=self.asset.device
-        ).unsqueeze(1)
-
-        # start from defaults and add random offsets
-        coms = self.default_com.clone()
-        coms[env_ids[:, None], body_ids, :3] += rand_samples
-
-        # Newton expects position-only (vec3f), PhysX expects the full pose (pos + quat)
-        # note: pass partial data of shape (len(env_ids), len(body_ids), ...) to match the API
-        if self._is_newton:
-            self.asset.set_coms_index(coms=coms[env_ids[:, None], body_ids, :3], body_ids=body_ids, env_ids=env_ids)
-        else:
-            self.asset.set_coms_index(coms=coms[env_ids[:, None], body_ids], body_ids=body_ids, env_ids=env_ids)
-
-
-class _RandomizeRigidBodyColliderOffsetsPhysx:
-    """PhysX backend implementation for collider offset randomization.
-
-    Uses rest offset and contact offset directly via the PhysX tensor API.
-    """
-
-    def __init__(self, asset: RigidObject | Articulation):
-        self.asset = asset
-        self.default_rest_offsets = wp.to_torch(asset.root_view.get_rest_offsets()).clone()
-        self.default_contact_offsets = wp.to_torch(asset.root_view.get_contact_offsets()).clone()
-
-    def __call__(
-        self,
-        env: ManagerBasedEnv,
-        env_ids: torch.Tensor | None,
-        asset_cfg: SceneEntityCfg,
-        rest_offset_distribution_params: tuple[float, float] | None = None,
-        contact_offset_distribution_params: tuple[float, float] | None = None,
-        distribution: Literal["uniform", "log_uniform", "gaussian"] = "uniform",
-    ):
-        if env_ids is None:
-            env_ids = torch.arange(env.scene.num_envs, device="cpu", dtype=torch.int32)
-        else:
-            env_ids = env_ids.to(device="cpu", dtype=torch.int32)
-        wp_env_ids = wp.from_torch(env_ids, dtype=wp.int32)
-
-        if rest_offset_distribution_params is not None:
-            rest_offset = self.default_rest_offsets.clone()
-            rest_offset = _randomize_prop_by_op(
-                rest_offset,
-                rest_offset_distribution_params,
-                None,
-                slice(None),
-                operation="abs",
-                distribution=distribution,
-            )
-            self.asset.root_view.set_rest_offsets(wp.from_torch(rest_offset), wp_env_ids)
-
-        if contact_offset_distribution_params is not None:
-            contact_offset = self.default_contact_offsets.clone()
-            contact_offset = _randomize_prop_by_op(
-                contact_offset,
-                contact_offset_distribution_params,
-                None,
-                slice(None),
-                operation="abs",
-                distribution=distribution,
-            )
-            self.asset.root_view.set_contact_offsets(wp.from_torch(contact_offset), wp_env_ids)
-
-
-class _RandomizeRigidBodyColliderOffsetsNewton:
-    """Newton backend implementation for collider offset randomization.
-
-    Maps PhysX concepts to Newton's geometry properties:
-
-    - ``rest_offset`` -> ``shape_margin`` (Newton margin)
-    - ``contact_offset`` -> ``shape_gap`` (Newton gap = contact_offset - margin)
-
-    See the `Newton collision schema`_ for details.
-
-    .. _Newton collision schema: https://newton-physics.github.io/newton/latest/concepts/collisions.html
-    """
-
-    def __init__(self, asset: RigidObject | Articulation):
-        import isaaclab_newton.physics.newton_manager as newton_manager_module  # noqa: PLC0415
-        from newton.solvers import SolverNotifyFlags  # noqa: PLC0415
-
-        self.asset = asset
-        self._newton_manager = newton_manager_module.NewtonManager
-        self._notify_shape_properties = SolverNotifyFlags.SHAPE_PROPERTIES
-
-        model = self._newton_manager.get_model()
-        self._sim_bind_shape_margin = asset._root_view.get_attribute("shape_margin", model)[:, 0]  # type: ignore
-        self._sim_bind_shape_gap = asset._root_view.get_attribute("shape_gap", model)[:, 0]  # type: ignore
-
-        self.default_margin = wp.to_torch(self._sim_bind_shape_margin).clone()
-        self.default_gap = wp.to_torch(self._sim_bind_shape_gap).clone()
-
-    def __call__(
-        self,
-        env: ManagerBasedEnv,
-        env_ids: torch.Tensor | None,
-        asset_cfg: SceneEntityCfg,
-        rest_offset_distribution_params: tuple[float, float] | None = None,
-        contact_offset_distribution_params: tuple[float, float] | None = None,
-        distribution: Literal["uniform", "log_uniform", "gaussian"] = "uniform",
-    ):
-        device = env.device
-        if env_ids is None:
-            env_ids = torch.arange(env.scene.num_envs, device=device, dtype=torch.int32)
-        else:
-            env_ids = env_ids.to(device)
-
-        margin_view = wp.to_torch(self._sim_bind_shape_margin)
-
-        if rest_offset_distribution_params is not None:
-            margin = self.default_margin.clone()
-            margin = _randomize_prop_by_op(
-                margin,
-                rest_offset_distribution_params,
-                None,
-                slice(None),
-                operation="abs",
-                distribution=distribution,
-            )
-            self.default_margin[env_ids] = margin[env_ids]
-            margin_view[env_ids] = margin[env_ids]
-
-        if contact_offset_distribution_params is not None:
-            current_margin = self.default_margin
-            contact_offset = torch.zeros_like(self.default_gap)
-            contact_offset = _randomize_prop_by_op(
-                contact_offset,
-                contact_offset_distribution_params,
-                None,
-                slice(None),
-                operation="abs",
-                distribution=distribution,
-            )
-            gap = torch.clamp(contact_offset - current_margin, min=0.0)
-            self.default_gap[env_ids] = gap[env_ids]
-            gap_view = wp.to_torch(self._sim_bind_shape_gap)
-            gap_view[env_ids] = gap[env_ids]
-
-        if rest_offset_distribution_params is not None or contact_offset_distribution_params is not None:
-            self._newton_manager.add_model_change(self._notify_shape_properties)
-
-
-class randomize_rigid_body_collider_offsets(ManagerTermBase):
-    """Randomize the collider parameters of rigid bodies by setting random values.
-
-    This function allows randomizing the collider parameters of the asset, such as rest and contact offsets.
-    These correspond to the physics engine collider properties that affect collision checking.
-
-    Automatically detects the active physics backend (PhysX or Newton) and delegates to
-    the appropriate backend-specific implementation:
-
-    - **PhysX**: Uses rest offset and contact offset directly via the PhysX tensor API
-      (``root_view.set_rest_offsets`` / ``root_view.set_contact_offsets``).
-    - **Newton**: Maps PhysX concepts to Newton's geometry properties. PhysX ``rest_offset``
-      maps to Newton ``shape_margin``, and PhysX ``contact_offset`` is converted to Newton
-      ``shape_gap`` via ``gap = contact_offset - margin``.
-
-    The function samples random values from the given distribution parameters and applies them
-    as absolute values to the collider properties. If the distribution parameters are not
-    provided for a particular property, the function does not modify it.
-
-    .. tip::
-        This function uses CPU tensors (PhysX) or GPU tensors (Newton) to assign the collision
-        properties. It is recommended to use this function only during the initialization of
-        the environment.
-    """
-
-    def __init__(self, cfg: EventTermCfg, env: ManagerBasedEnv):
-        """Initialize the term.
-
-        Args:
-            cfg: The configuration of the event term.
-            env: The environment instance.
-
-        Raises:
-            ValueError: If the asset is not a RigidObject or an Articulation.
-        """
-        from isaaclab.assets import BaseArticulation, BaseRigidObject
-
-        super().__init__(cfg, env)
-
-        self.asset_cfg: SceneEntityCfg = cfg.params["asset_cfg"]
-        self.asset: RigidObject | Articulation = env.scene[self.asset_cfg.name]
-
-        if not isinstance(self.asset, (BaseRigidObject, BaseArticulation)):
-            raise ValueError(
-                f"Randomization term 'randomize_rigid_body_collider_offsets' not supported for asset:"
-                f" '{self.asset_cfg.name}' with type: '{type(self.asset)}'."
-            )
-
-        # detect physics backend and instantiate the appropriate implementation
-        manager_name = env.sim.physics_manager.__name__.lower()
-        if "newton" in manager_name:
-            self._impl = _RandomizeRigidBodyColliderOffsetsNewton(self.asset)
-        else:
-            self._impl = _RandomizeRigidBodyColliderOffsetsPhysx(self.asset)
-
-    def __call__(
-        self,
-        env: ManagerBasedEnv,
-        env_ids: torch.Tensor | None,
-        asset_cfg: SceneEntityCfg,
-        rest_offset_distribution_params: tuple[float, float] | None = None,
-        contact_offset_distribution_params: tuple[float, float] | None = None,
-        distribution: Literal["uniform", "log_uniform", "gaussian"] = "uniform",
-    ):
-        self._impl(
-            env,
-            env_ids,
-            asset_cfg,
-            rest_offset_distribution_params,
-            contact_offset_distribution_params,
-            distribution,
-        )
-
-
-class randomize_physics_scene_gravity(ManagerTermBase):
-    """Randomize gravity by adding, scaling, or setting random values.
-
-    Automatically detects the active physics backend (PhysX or Newton) and applies
-    the appropriate gravity randomization strategy:
-
-    - **PhysX**: samples a single gravity vector and sets it scene-wide via the PhysX
-      simulation view.  All environments share the same gravity.
-    - **Newton**: samples per-environment gravity vectors and writes them in-place to
-      the Newton model's per-world gravity array on GPU.
-
-    The distribution parameters are tuples of two lists with three floats each,
-    representing the lower and upper bounds for the x, y, and z gravity components [m/s^2].
-
-    Args:
-        cfg: The configuration of the event term.
-        env: The environment instance.
-    """
-
-    def __init__(self, cfg: EventTermCfg, env: ManagerBasedEnv):
-        super().__init__(cfg, env)
-
-        manager_name = env.sim.physics_manager.__name__.lower()
-        if "newton" in manager_name:
-            self._backend = "newton"
-            self._init_newton(cfg, env)
-        else:
-            self._backend = "physx"
-            self._init_physx(env)
-
-        distribution = cfg.params.get("distribution", "uniform")
-        if distribution == "uniform":
-            self._dist_fn = math_utils.sample_uniform
-        elif distribution == "log_uniform":
-            self._dist_fn = math_utils.sample_log_uniform
-        elif distribution == "gaussian":
-            self._dist_fn = math_utils.sample_gaussian
-        else:
-            raise NotImplementedError(
-                f"Unknown distribution: '{distribution}' for gravity randomization."
-                " Please use 'uniform', 'log_uniform', or 'gaussian'."
-            )
-
-        operation = cfg.params["operation"]
-        if operation not in ("add", "scale", "abs"):
-            raise NotImplementedError(
-                f"Unknown operation: '{operation}' for gravity randomization. Please use 'add', 'scale', or 'abs'."
-            )
-
-        gravity_distribution_params = cfg.params["gravity_distribution_params"]
-        self._dist_param_0 = torch.tensor(gravity_distribution_params[0], device=env.device, dtype=torch.float32)
-        self._dist_param_1 = torch.tensor(gravity_distribution_params[1], device=env.device, dtype=torch.float32)
-
-    def __call__(
-        self,
-        env: ManagerBasedEnv,
-        env_ids: torch.Tensor | None,
-        gravity_distribution_params: tuple[list[float], list[float]],
-        operation: Literal["add", "scale", "abs"],
-        distribution: Literal["uniform", "log_uniform", "gaussian"] = "uniform",
-    ):
-        """Randomize gravity for the specified environments.
-
-        Args:
-            env: The environment instance.
-            env_ids: The environment IDs to randomize. If None, all environments are randomized.
-            gravity_distribution_params: Distribution parameters as a tuple of two lists, each
-                with 3 floats corresponding to (x, y, z) gravity components [m/s^2]. Updated
-                into pre-allocated tensors each call to support curriculum-driven range changes.
-            operation: The operation to apply ('add', 'scale', or 'abs').
-            distribution: The distribution type (cached at init, param ignored at runtime).
-        """
-        self._dist_param_0[0] = gravity_distribution_params[0][0]
-        self._dist_param_1[0] = gravity_distribution_params[1][0]
-        self._dist_param_0[1] = gravity_distribution_params[0][1]
-        self._dist_param_1[1] = gravity_distribution_params[1][1]
-        self._dist_param_0[2] = gravity_distribution_params[0][2]
-        self._dist_param_1[2] = gravity_distribution_params[1][2]
-
-        if self._backend == "newton":
-            self._call_newton(env, env_ids, operation)
-        else:
-            self._call_physx(env, operation)
-
-    def _init_newton(self, cfg: EventTermCfg, env: ManagerBasedEnv):
-        """Cache Newton manager reference and solver notification flag."""
-        import isaaclab_newton.physics.newton_manager as newton_manager_module  # noqa: PLC0415
-        from newton.solvers import SolverNotifyFlags  # noqa: PLC0415
-
-        self._newton_manager = newton_manager_module.NewtonManager
-        self._notify_model_properties = SolverNotifyFlags.MODEL_PROPERTIES
-
-    def _call_newton(
-        self,
-        env: ManagerBasedEnv,
-        env_ids: torch.Tensor | None,
-        operation: str,
-    ):
-        """Apply per-environment gravity via Newton's per-world gravity array on GPU."""
-        model = self._newton_manager.get_model()
-        if model is None or model.gravity is None:
-            raise RuntimeError("Newton model is not initialized. Cannot randomize gravity.")
-
-        gravity = wp.to_torch(model.gravity)
-
-        if env_ids is None:
-            env_ids = env.scene._ALL_INDICES
-        if len(env_ids) == 0:
-            return
-
-        num = len(env_ids)
-        random_values = self._dist_fn(
-            self._dist_param_0.unsqueeze(0).expand(num, -1),
-            self._dist_param_1.unsqueeze(0).expand(num, -1),
-            (num, 3),
-            device=env.device,
-        )
-
-        if operation == "abs":
-            gravity[env_ids] = random_values
-        elif operation == "add":
-            gravity[env_ids] += random_values
-        elif operation == "scale":
-            gravity[env_ids] *= random_values
-
-        self._newton_manager.add_model_change(self._notify_model_properties)
-
-    def _init_physx(self, env: ManagerBasedEnv):
-        """Cache the ``carb`` module and PhysX simulation view for scene-wide gravity updates."""
-        import carb  # noqa: PLC0415
-
-        self._carb = carb
-        self._physics_sim_view = sim_utils.SimulationContext.instance().physics_sim_view
-
-    def _call_physx(self, env: ManagerBasedEnv, operation: str):
-        """Sample a single gravity vector and apply it scene-wide via the PhysX simulation view."""
-        gravity = torch.tensor(env.sim.cfg.gravity, device="cpu").unsqueeze(0)
-        gravity = _randomize_prop_by_op(
-            gravity,
-            (self._dist_param_0.cpu(), self._dist_param_1.cpu()),
-            None,
-            slice(None),
-            operation=operation,
-            distribution="uniform",
-        )
-        gravity = gravity[0].tolist()
-        self._physics_sim_view.set_gravity(self._carb.Float3(*gravity))
-
-
-class randomize_actuator_gains(ManagerTermBase):
-    """Randomize the actuator gains in an articulation by adding, scaling, or setting random values.
-
-    This function allows randomizing the actuator stiffness and damping gains.
-
-    The function samples random values from the given distribution parameters and applies the operation to
-    the joint properties. It then sets the values into the actuator models. If the distribution parameters
-    are not provided for a particular property, the function does not modify the property.
-
-    .. tip::
-        For implicit actuators, this function uses CPU tensors to assign the actuator gains into the simulation.
-        In such cases, it is recommended to use this function only during the initialization of the environment.
-    """
-
-    def __init__(self, cfg: EventTermCfg, env: ManagerBasedEnv):
-        """Initialize the term.
-
-        Args:
-            cfg: The configuration of the event term.
-            env: The environment instance.
-
-        Raises:
-            TypeError: If `params` is not a tuple of two numbers.
-            ValueError: If the operation is not supported.
-            ValueError: If the lower bound is negative or zero when not allowed.
-            ValueError: If the upper bound is less than the lower bound.
-        """
-        super().__init__(cfg, env)
-
-        # extract the used quantities (to enable type-hinting)
-        self.asset_cfg: SceneEntityCfg = cfg.params["asset_cfg"]
-        self.asset: RigidObject | Articulation = env.scene[self.asset_cfg.name]
-
-        self.default_joint_stiffness = self.asset.data.joint_stiffness.torch.clone()
-        self.default_joint_damping = self.asset.data.joint_damping.torch.clone()
-
-        # For explicit Lab actuators the sim-level stiffness/damping is zeroed out,
-        # so patch the defaults with the actual actuator PD gains.
-        for actuator in self.asset.actuators.values():
-            if not isinstance(actuator, ImplicitActuator):
-                joint_ids = actuator.joint_indices
-                self.default_joint_stiffness[:, joint_ids] = actuator.stiffness
-                self.default_joint_damping[:, joint_ids] = actuator.damping
-        # Same for explicit Newton actuators on either backend — their kp/kd
-        # live on the per-actuator controller arrays (not on a Lab Actuator
-        # object), so the asset exposes a per-articulation snapshot taken
-        # at articulation init time.
-        newton_default_stiffness = getattr(self.asset, "newton_default_stiffness", None)
-        if newton_default_stiffness is not None:
-            joint_ids = self.asset.newton_managed_local_joints
-            self.default_joint_stiffness[:, joint_ids] = newton_default_stiffness[:, joint_ids]
-            self.default_joint_damping[:, joint_ids] = self.asset.newton_default_damping[:, joint_ids]
-
-        # check for valid operation
-        if cfg.params["operation"] == "scale":
-            if "stiffness_distribution_params" in cfg.params:
-                _validate_scale_range(
-                    cfg.params["stiffness_distribution_params"], "stiffness_distribution_params", allow_zero=False
-                )
-            if "damping_distribution_params" in cfg.params:
-                _validate_scale_range(cfg.params["damping_distribution_params"], "damping_distribution_params")
-        elif cfg.params["operation"] not in ("abs", "add"):
-            raise ValueError(
-                "Randomization term 'randomize_actuator_gains' does not support operation:"
-                f" '{cfg.params['operation']}'."
-            )
-
-    def __call__(
-        self,
-        env: ManagerBasedEnv,
-        env_ids: torch.Tensor | None,
-        asset_cfg: SceneEntityCfg,
-        stiffness_distribution_params: tuple[float, float] | None = None,
-        damping_distribution_params: tuple[float, float] | None = None,
-        operation: Literal["add", "scale", "abs"] = "abs",
-        distribution: Literal["uniform", "log_uniform", "gaussian"] = "uniform",
-    ):
-        # Resolve environment ids
-        if env_ids is None:
-            env_ids = torch.arange(env.scene.num_envs, device=self.asset.device)
-
-        def randomize(data: torch.Tensor, params: tuple[float, float]) -> torch.Tensor:
-            return _randomize_prop_by_op(
-                data, params, dim_0_ids=None, dim_1_ids=actuator_indices, operation=operation, distribution=distribution
-            )
-
-        # Loop through actuators and randomize gains
-        for actuator in self.asset.actuators.values():
-            if isinstance(self.asset_cfg.joint_ids, slice):
-                # we take all the joints of the actuator
-                actuator_indices = slice(None)
-                if isinstance(actuator.joint_indices, slice):
-                    global_indices = slice(None)
-                elif isinstance(actuator.joint_indices, torch.Tensor):
-                    global_indices = actuator.joint_indices.to(self.asset.device)
-                else:
-                    raise TypeError("Actuator joint indices must be a slice or a torch.Tensor.")
-            elif isinstance(actuator.joint_indices, slice):
-                # we take the joints defined in the asset config
-                global_indices = actuator_indices = torch.tensor(self.asset_cfg.joint_ids, device=self.asset.device)
-            else:
-                # we take the intersection of the actuator joints and the asset config joints
-                actuator_joint_indices = actuator.joint_indices
-                asset_joint_ids = torch.tensor(self.asset_cfg.joint_ids, device=self.asset.device)
-                # the indices of the joints in the actuator that have to be randomized
-                actuator_indices = torch.nonzero(torch.isin(actuator_joint_indices, asset_joint_ids)).view(-1)
-                if len(actuator_indices) == 0:
-                    continue
-                # maps actuator indices that have to be randomized to global joint indices
-                global_indices = actuator_joint_indices[actuator_indices]
-            # Randomize stiffness
-            if stiffness_distribution_params is not None:
-                stiffness = actuator.stiffness[env_ids].clone()
-                stiffness[:, actuator_indices] = self.default_joint_stiffness[env_ids][:, global_indices].clone()
-                randomize(stiffness, stiffness_distribution_params)
-                actuator.stiffness[env_ids] = stiffness
-                if isinstance(actuator, ImplicitActuator):
-                    self.asset.write_joint_stiffness_to_sim_index(
-                        stiffness=stiffness, joint_ids=actuator.joint_indices, env_ids=env_ids
-                    )
-            # Randomize damping
-            if damping_distribution_params is not None:
-                damping = actuator.damping[env_ids].clone()
-                damping[:, actuator_indices] = self.default_joint_damping[env_ids][:, global_indices].clone()
-                randomize(damping, damping_distribution_params)
-                actuator.damping[env_ids] = damping
-                if isinstance(actuator, ImplicitActuator):
-                    self.asset.write_joint_damping_to_sim_index(
-                        damping=damping, joint_ids=actuator.joint_indices, env_ids=env_ids
-                    )
-
-        # Push DR updates to explicit Newton-actuator controllers via the asset's
-        # own write methods. Each backend's articulation iterates the adapter's
-        # actuators and propagates per actuator, using the appropriate backend
-        # mechanism (Newton ``ArticulationView`` on the Newton backend, an
-        # in-place scatter kernel on PhysX).
-        if not hasattr(self.asset, "write_actuator_stiffness_to_sim"):
-            return
-
-        if isinstance(self.asset_cfg.joint_ids, slice):
-            joint_ids = torch.arange(self.asset.num_joints, device=self.asset.device, dtype=torch.long)
-        else:
-            joint_ids = torch.tensor(self.asset_cfg.joint_ids, device=self.asset.device, dtype=torch.long)
-
-        if stiffness_distribution_params is not None:
-            new_stiffness = self.default_joint_stiffness[env_ids][:, joint_ids].clone()
-            _randomize_prop_by_op(
-                new_stiffness,
-                stiffness_distribution_params,
-                dim_0_ids=None,
-                dim_1_ids=slice(None),
-                operation=operation,
-                distribution=distribution,
-            )
-            self.asset.write_actuator_stiffness_to_sim(
-                stiffness=new_stiffness,
-                env_ids=env_ids,
-                joint_ids=joint_ids,
-            )
-        if damping_distribution_params is not None:
-            new_damping = self.default_joint_damping[env_ids][:, joint_ids].clone()
-            _randomize_prop_by_op(
-                new_damping,
-                damping_distribution_params,
-                dim_0_ids=None,
-                dim_1_ids=slice(None),
-                operation=operation,
-                distribution=distribution,
-            )
-            self.asset.write_actuator_damping_to_sim(
-                damping=new_damping,
-                env_ids=env_ids,
-                joint_ids=joint_ids,
-            )
-
-
-class randomize_joint_parameters(ManagerTermBase):
-    """Randomize the simulated joint parameters of an articulation by adding, scaling, or setting random values.
-
-    This function allows randomizing the joint parameters of the asset. These correspond to the physics engine
-    joint properties that affect the joint behavior. The properties include the joint friction coefficient, armature,
-    and joint position limits.
-
-    The function samples random values from the given distribution parameters and applies the operation to the
-    joint properties. It then sets the values into the physics simulation. If the distribution parameters are
-    not provided for a particular property, the function does not modify the property.
-
-    .. tip::
-        This function uses CPU tensors to assign the joint properties. It is recommended to use this function
-        only during the initialization of the environment.
-    """
-
-    def __init__(self, cfg: EventTermCfg, env: ManagerBasedEnv):
-        """Initialize the term.
-
-        Args:
-            cfg: The configuration of the event term.
-            env: The environment instance.
-
-        Raises:
-            TypeError: If `params` is not a tuple of two numbers.
-            ValueError: If the operation is not supported.
-            ValueError: If the lower bound is negative or zero when not allowed.
-            ValueError: If the upper bound is less than the lower bound.
-        """
-        super().__init__(cfg, env)
-
-        # extract the used quantities (to enable type-hinting)
-        self.asset_cfg: SceneEntityCfg = cfg.params["asset_cfg"]
-        self.asset: Articulation = env.scene[self.asset_cfg.name]
-
-        # detect physics backend
-        manager_name = env.sim.physics_manager.__name__.lower()
-        self._backend = "newton" if "newton" in manager_name else "physx"
-
-        # cache default values (common to both backends)
-        self.default_joint_friction_coeff = self.asset.data.joint_friction_coeff.torch.clone()
-        self.default_joint_armature = self.asset.data.joint_armature.torch.clone()
-        self.default_joint_pos_limits = self.asset.data.joint_pos_limits.torch.clone()
-
-        # cache dynamic/viscous friction (PhysX only - Newton only has static friction)
-        if self._backend == "physx":
-            self.default_dynamic_joint_friction_coeff = (self.asset.data.joint_dynamic_friction_coeff.torch).clone()
-            self.default_viscous_joint_friction_coeff = (self.asset.data.joint_viscous_friction_coeff.torch).clone()
-
-        # check for valid operation
-        if cfg.params["operation"] == "scale":
-            if "friction_distribution_params" in cfg.params:
-                _validate_scale_range(cfg.params["friction_distribution_params"], "friction_distribution_params")
-            if "armature_distribution_params" in cfg.params:
-                _validate_scale_range(cfg.params["armature_distribution_params"], "armature_distribution_params")
-        elif cfg.params["operation"] not in ("abs", "add"):
-            raise ValueError(
-                "Randomization term 'randomize_fixed_tendon_parameters' does not support operation:"
-                f" '{cfg.params['operation']}'."
-            )
-
-    def __call__(
-        self,
-        env: ManagerBasedEnv,
-        env_ids: torch.Tensor | None,
-        asset_cfg: SceneEntityCfg,
-        friction_distribution_params: tuple[float, float] | None = None,
-        armature_distribution_params: tuple[float, float] | None = None,
-        lower_limit_distribution_params: tuple[float, float] | None = None,
-        upper_limit_distribution_params: tuple[float, float] | None = None,
-        operation: Literal["add", "scale", "abs"] = "abs",
-        distribution: Literal["uniform", "log_uniform", "gaussian"] = "uniform",
-    ):
-        # resolve environment ids
-        if env_ids is None:
-            env_ids = torch.arange(env.scene.num_envs, device=self.asset.device)
-
-        # resolve joint indices
-        if self.asset_cfg.joint_ids == slice(None):
-            joint_ids = slice(None)  # for optimization purposes
-        else:
-            joint_ids = torch.tensor(self.asset_cfg.joint_ids, dtype=torch.int, device=self.asset.device)
-
-        if env_ids != slice(None) and joint_ids != slice(None):
-            env_ids_for_slice = env_ids[:, None]
-        else:
-            env_ids_for_slice = env_ids
-
-        # sample joint properties from the given ranges and set into the physics simulation
-        # joint friction coefficient
-        if friction_distribution_params is not None:
-            friction_coeff = _randomize_prop_by_op(
-                self.default_joint_friction_coeff.clone(),
-                friction_distribution_params,
-                env_ids,
-                joint_ids,
-                operation=operation,
-                distribution=distribution,
-            )
-
-            # ensure the friction coefficient is non-negative
-            friction_coeff = torch.clamp(friction_coeff, min=0.0)
-
-            # Always set static friction (indexed once)
-            static_friction_coeff = friction_coeff[env_ids_for_slice, joint_ids]
-
-            if self._backend == "newton":
-                # Newton only supports static friction coefficient
-                self.asset.write_joint_friction_coefficient_to_sim_index(
-                    joint_friction_coeff=static_friction_coeff,
-                    joint_ids=joint_ids,
-                    env_ids=env_ids,
-                )
-            else:
-                # Randomize raw tensors
-                dynamic_friction_coeff = _randomize_prop_by_op(
-                    self.default_dynamic_joint_friction_coeff.clone(),
-                    friction_distribution_params,
-                    env_ids,
-                    joint_ids,
-                    operation=operation,
-                    distribution=distribution,
-                )
-                viscous_friction_coeff = _randomize_prop_by_op(
-                    self.default_viscous_joint_friction_coeff.clone(),
-                    friction_distribution_params,
-                    env_ids,
-                    joint_ids,
-                    operation=operation,
-                    distribution=distribution,
-                )
-
-                # Clamp to non-negative
-                dynamic_friction_coeff = torch.clamp(dynamic_friction_coeff, min=0.0)
-                viscous_friction_coeff = torch.clamp(viscous_friction_coeff, min=0.0)
-
-                # Ensure dynamic ≤ static (same shape before indexing)
-                dynamic_friction_coeff = torch.minimum(dynamic_friction_coeff, friction_coeff)
-
-                # Index once at the end
-                dynamic_friction_coeff = dynamic_friction_coeff[env_ids_for_slice, joint_ids]
-                viscous_friction_coeff = viscous_friction_coeff[env_ids_for_slice, joint_ids]
-
-                # Single write call for all versions
-                self.asset.write_joint_friction_coefficient_to_sim_index(
-                    joint_friction_coeff=static_friction_coeff,
-                    joint_dynamic_friction_coeff=dynamic_friction_coeff,
-                    joint_viscous_friction_coeff=viscous_friction_coeff,
-                    joint_ids=joint_ids,
-                    env_ids=env_ids,
-                )
-
-        # joint armature
-        if armature_distribution_params is not None:
-            armature = _randomize_prop_by_op(
-                self.asset.data.default_joint_armature.torch.clone(),
-                armature_distribution_params,
-                env_ids,
-                joint_ids,
-                operation=operation,
-                distribution=distribution,
-            )
-            self.asset.write_joint_armature_to_sim(
-                armature[env_ids_for_slice, joint_ids], joint_ids=joint_ids, env_ids=env_ids
-            )
-
-        # joint position limits
-        if lower_limit_distribution_params is not None or upper_limit_distribution_params is not None:
-            joint_pos_limits = self.default_joint_pos_limits.clone()
-            # -- randomize the lower limits
-            if lower_limit_distribution_params is not None:
-                joint_pos_limits[..., 0] = _randomize_prop_by_op(
-                    joint_pos_limits[..., 0],
-                    lower_limit_distribution_params,
-                    env_ids,
-                    joint_ids,
-                    operation=operation,
-                    distribution=distribution,
-                )
-            # -- randomize the upper limits
-            if upper_limit_distribution_params is not None:
-                joint_pos_limits[..., 1] = _randomize_prop_by_op(
-                    joint_pos_limits[..., 1],
-                    upper_limit_distribution_params,
-                    env_ids,
-                    joint_ids,
-                    operation=operation,
-                    distribution=distribution,
-                )
-
-            # extract the position limits for the concerned joints
-            joint_pos_limits = joint_pos_limits[env_ids_for_slice, joint_ids]
-            if (joint_pos_limits[..., 0] > joint_pos_limits[..., 1]).any():
-                raise ValueError(
-                    "Randomization term 'randomize_joint_parameters' is setting lower joint limits that are greater"
-                    " than upper joint limits. Please check the distribution parameters for the joint position limits."
-                )
-            # set the position limits into the physics simulation
-            self.asset.write_joint_position_limit_to_sim_index(
-                limits=joint_pos_limits, joint_ids=joint_ids, env_ids=env_ids, warn_limit_violation=False
-            )
-
-
-class randomize_fixed_tendon_parameters(ManagerTermBase):
-    """Randomize the simulated fixed tendon parameters of an articulation by adding, scaling, or setting random values.
-
-    This function allows randomizing the fixed tendon parameters of the asset.
-    These correspond to the physics engine tendon properties that affect the joint behavior.
-
-    The function samples random values from the given distribution parameters and applies the operation to
-    the tendon properties. It then sets the values into the physics simulation. If the distribution parameters
-    are not provided for a particular property, the function does not modify the property.
-    """
-
-    def __init__(self, cfg: EventTermCfg, env: ManagerBasedEnv):
-        """Initialize the term.
-
-        Args:
-            cfg: The configuration of the event term.
-            env: The environment instance.
-
-        Raises:
-            TypeError: If `params` is not a tuple of two numbers.
-            ValueError: If the operation is not supported.
-            ValueError: If the lower bound is negative or zero when not allowed.
-            ValueError: If the upper bound is less than the lower bound.
-        """
-        super().__init__(cfg, env)
-
-        # extract the used quantities (to enable type-hinting)
-        self.asset_cfg: SceneEntityCfg = cfg.params["asset_cfg"]
-        self.asset: RigidObject | Articulation = env.scene[self.asset_cfg.name]
-        # check for valid operation
-        if cfg.params["operation"] == "scale":
-            if "stiffness_distribution_params" in cfg.params:
-                _validate_scale_range(
-                    cfg.params["stiffness_distribution_params"], "stiffness_distribution_params", allow_zero=False
-                )
-            if "damping_distribution_params" in cfg.params:
-                _validate_scale_range(cfg.params["damping_distribution_params"], "damping_distribution_params")
-            if "limit_stiffness_distribution_params" in cfg.params:
-                _validate_scale_range(
-                    cfg.params["limit_stiffness_distribution_params"], "limit_stiffness_distribution_params"
-                )
-        elif cfg.params["operation"] not in ("abs", "add"):
-            raise ValueError(
-                "Randomization term 'randomize_fixed_tendon_parameters' does not support operation:"
-                f" '{cfg.params['operation']}'."
-            )
-
-    def __call__(
-        self,
-        env: ManagerBasedEnv,
-        env_ids: torch.Tensor | None,
-        asset_cfg: SceneEntityCfg,
-        stiffness_distribution_params: tuple[float, float] | None = None,
-        damping_distribution_params: tuple[float, float] | None = None,
-        limit_stiffness_distribution_params: tuple[float, float] | None = None,
-        lower_limit_distribution_params: tuple[float, float] | None = None,
-        upper_limit_distribution_params: tuple[float, float] | None = None,
-        rest_length_distribution_params: tuple[float, float] | None = None,
-        offset_distribution_params: tuple[float, float] | None = None,
-        operation: Literal["add", "scale", "abs"] = "abs",
-        distribution: Literal["uniform", "log_uniform", "gaussian"] = "uniform",
-    ):
-        _backend = env.sim.physics_manager.__name__.lower()
-
-        # resolve environment ids
-        if env_ids is None:
-            env_ids = torch.arange(env.scene.num_envs, device=self.asset.device)
-
-        # resolve joint indices
-        if self.asset_cfg.fixed_tendon_ids == slice(None):
-            tendon_ids = slice(None)  # for optimization purposes
-        else:
-            tendon_ids = torch.tensor(self.asset_cfg.fixed_tendon_ids, dtype=torch.int, device=self.asset.device)
-
-        # sample tendon properties from the given ranges and set into the physics simulation
-        # stiffness
-        if stiffness_distribution_params is not None:
-            stiffness = _randomize_prop_by_op(
-                self.asset.data.fixed_tendon_stiffness.torch.clone(),
-                stiffness_distribution_params,
-                env_ids,
-                tendon_ids,
-                operation=operation,
-                distribution=distribution,
-            )
-            self.asset.set_fixed_tendon_stiffness_index(
-                stiffness=stiffness[env_ids[:, None], tendon_ids], fixed_tendon_ids=tendon_ids, env_ids=env_ids
-            )
-
-        # damping
-        if damping_distribution_params is not None:
-            damping = _randomize_prop_by_op(
-                self.asset.data.fixed_tendon_damping.torch.clone(),
-                damping_distribution_params,
-                env_ids,
-                tendon_ids,
-                operation=operation,
-                distribution=distribution,
-            )
-            self.asset.set_fixed_tendon_damping_index(
-                damping=damping[env_ids[:, None], tendon_ids], fixed_tendon_ids=tendon_ids, env_ids=env_ids
-            )
-
-        # limit stiffness
-        if limit_stiffness_distribution_params is not None:
-            if _backend == "physx":
-                limit_stiffness = _randomize_prop_by_op(
-                    self.asset.data.fixed_tendon_limit_stiffness.torch.clone(),
-                    limit_stiffness_distribution_params,
-                    env_ids,
-                    tendon_ids,
-                    operation=operation,
-                    distribution=distribution,
-                )
-                self.asset.set_fixed_tendon_limit_stiffness(
-                    limit_stiffness[env_ids[:, None], tendon_ids], tendon_ids, env_ids
-                )
-            else:
-                raise NotImplementedError("Limit stiffness is not support in Newton.")
-
-        # position limits
-        if lower_limit_distribution_params is not None or upper_limit_distribution_params is not None:
-            if _backend == "physx":
-                limit = self.asset.data.fixed_tendon_pos_limits.torch.clone()
-                # -- lower limit
-                if lower_limit_distribution_params is not None:
-                    limit[..., 0] = _randomize_prop_by_op(
-                        limit[..., 0],
-                        lower_limit_distribution_params,
-                        env_ids,
-                        tendon_ids,
-                        operation=operation,
-                        distribution=distribution,
-                    )
-                # -- upper limit
-                if upper_limit_distribution_params is not None:
-                    limit[..., 1] = _randomize_prop_by_op(
-                        limit[..., 1],
-                        upper_limit_distribution_params,
-                        env_ids,
-                        tendon_ids,
-                        operation=operation,
-                        distribution=distribution,
-                    )
-
-                # check if the limits are valid
-                tendon_limits = limit[env_ids[:, None], tendon_ids]
-                if (tendon_limits[..., 0] > tendon_limits[..., 1]).any():
-                    raise ValueError(
-                        "Randomization term 'randomize_fixed_tendon_parameters' is setting lower tendon limits that are"
-                        " greater than upper tendon limits."
-                    )
-                self.asset.set_fixed_tendon_position_limit_index(
-                    limit=tendon_limits, fixed_tendon_ids=tendon_ids, env_ids=env_ids
-                )
-            else:
-                raise NotImplementedError("Position limits is not yet implemented with Newton.")
-
-        # rest length
-        if rest_length_distribution_params is not None:
-            if _backend == "physx":
-                rest_length = _randomize_prop_by_op(
-                    self.asset.data.fixed_tendon_rest_length.torch.clone(),
-                    rest_length_distribution_params,
-                    env_ids,
-                    tendon_ids,
-                    operation=operation,
-                    distribution=distribution,
-                )
-                self.asset.set_fixed_tendon_rest_length_index(
-                    rest_length=rest_length[env_ids[:, None], tendon_ids], fixed_tendon_ids=tendon_ids, env_ids=env_ids
-                )
-            else:
-                raise NotImplementedError("Rest length is not yet implemented with Newton.")
-        # offset
-        if offset_distribution_params is not None:
-            if _backend == "physx":
-                offset = _randomize_prop_by_op(
-                    self.asset.data.fixed_tendon_offset.torch.clone(),
-                    offset_distribution_params,
-                    env_ids,
-                    tendon_ids,
-                    operation=operation,
-                    distribution=distribution,
-                )
-                self.asset.set_fixed_tendon_offset_index(
-                    offset=offset[env_ids[:, None], tendon_ids], fixed_tendon_ids=tendon_ids, env_ids=env_ids
-                )
-            else:
-                raise NotImplementedError("Offset is not supported in Newton.")
-
-        # write the fixed tendon properties into the simulation
-        self.asset.write_fixed_tendon_properties_to_sim_index(env_ids=env_ids)
 
 
 def apply_external_force_torque(
@@ -2436,112 +775,3 @@ class randomize_visual_color(ManagerTermBase):
 """
 Internal helper functions.
 """
-
-
-def _randomize_prop_by_op(
-    data: torch.Tensor,
-    distribution_parameters: tuple[float | torch.Tensor, float | torch.Tensor],
-    dim_0_ids: torch.Tensor | None,
-    dim_1_ids: torch.Tensor | slice,
-    operation: Literal["add", "scale", "abs"],
-    distribution: Literal["uniform", "log_uniform", "gaussian"],
-) -> torch.Tensor:
-    """Perform data randomization based on the given operation and distribution.
-
-    Args:
-        data: The data tensor to be randomized. Shape is (dim_0, dim_1).
-        distribution_parameters: The parameters for the distribution to sample values from.
-        dim_0_ids: The indices of the first dimension to randomize.
-        dim_1_ids: The indices of the second dimension to randomize.
-        operation: The operation to perform on the data. Options: 'add', 'scale', 'abs'.
-        distribution: The distribution to sample the random values from. Options: 'uniform', 'log_uniform'.
-
-    Returns:
-        The data tensor after randomization. Shape is (dim_0, dim_1).
-
-    Raises:
-        NotImplementedError: If the operation or distribution is not supported.
-    """
-    # resolve shape
-    # -- dim 0
-    if dim_0_ids is None:
-        n_dim_0 = data.shape[0]
-        dim_0_ids = slice(None)
-    else:
-        n_dim_0 = len(dim_0_ids)
-        if not isinstance(dim_1_ids, slice):
-            dim_0_ids = dim_0_ids[:, None]
-    # -- dim 1
-    if isinstance(dim_1_ids, slice):
-        n_dim_1 = data.shape[1]
-    else:
-        n_dim_1 = len(dim_1_ids)
-
-    # resolve the distribution
-    if distribution == "uniform":
-        dist_fn = math_utils.sample_uniform
-    elif distribution == "log_uniform":
-        dist_fn = math_utils.sample_log_uniform
-    elif distribution == "gaussian":
-        dist_fn = math_utils.sample_gaussian
-    else:
-        raise NotImplementedError(
-            f"Unknown distribution: '{distribution}' for joint properties randomization."
-            " Please use 'uniform', 'log_uniform', 'gaussian'."
-        )
-    # perform the operation
-    if operation == "add":
-        data[dim_0_ids, dim_1_ids] += dist_fn(*distribution_parameters, (n_dim_0, n_dim_1), device=data.device)
-    elif operation == "scale":
-        data[dim_0_ids, dim_1_ids] *= dist_fn(*distribution_parameters, (n_dim_0, n_dim_1), device=data.device)
-    elif operation == "abs":
-        data[dim_0_ids, dim_1_ids] = dist_fn(*distribution_parameters, (n_dim_0, n_dim_1), device=data.device)
-    else:
-        raise NotImplementedError(
-            f"Unknown operation: '{operation}' for property randomization. Please use 'add', 'scale', or 'abs'."
-        )
-    return data
-
-
-def _validate_scale_range(
-    params: tuple[float, float] | None,
-    name: str,
-    *,
-    allow_negative: bool = False,
-    allow_zero: bool = True,
-) -> None:
-    """
-    Validates a (low, high) tuple used in scale-based randomization.
-
-    This function ensures the tuple follows expected rules when applying a 'scale'
-    operation. It performs type and value checks, optionally allowing negative or
-    zero lower bounds.
-
-    Args:
-        params (tuple[float, float] | None): The (low, high) range to validate. If None,
-            validation is skipped.
-        name (str): The name of the parameter being validated, used for error messages.
-        allow_negative (bool, optional): If True, allows the lower bound to be negative.
-            Defaults to False.
-        allow_zero (bool, optional): If True, allows the lower bound to be zero.
-            Defaults to True.
-
-    Raises:
-        TypeError: If `params` is not a tuple of two numbers.
-        ValueError: If the lower bound is negative or zero when not allowed.
-        ValueError: If the upper bound is less than the lower bound.
-
-    Example:
-        _validate_scale_range((0.5, 1.5), "mass_scale")
-    """
-    if params is None:  # caller didn’t request randomisation for this field
-        return
-    low, high = params
-    if not isinstance(low, (int, float)) or not isinstance(high, (int, float)):
-        raise TypeError(f"{name}: expected (low, high) to be a tuple of numbers, got {params}.")
-    if not allow_negative and not allow_zero and low <= 0:
-        raise ValueError(f"{name}: lower bound must be > 0 when using the 'scale' operation (got {low}).")
-    if not allow_negative and allow_zero and low < 0:
-        raise ValueError(f"{name}: lower bound must be ≥ 0 when using the 'scale' operation (got {low}).")
-    if high < low:
-        raise ValueError(f"{name}: upper bound ({high}) must be ≥ lower bound ({low}).")
