+++ source/isaaclab_experimental/changelog.d/zhengyuz-decouple-physics-events.rst
@@ -1,0 +1,5 @@
+Changed
+^^^^^^^
+
+* Updated ``isaaclab_experimental.envs.mdp`` to forward the physics event terms from their new
+  location in :mod:`isaaclab.envs.mdp.physics_events` (re-exported via :mod:`isaaclab.envs.mdp`).
