+++ source/isaaclab_newton/changelog.d/zhengyuz-decouple-physics-events.rst
@@ -1,0 +1,6 @@
+Added
+^^^^^
+
+* Added Newton backend implementations of the MDP physics event-randomization terms under
+  ``isaaclab_newton.envs.mdp``. These are resolved at runtime by the converters in
+  :mod:`isaaclab.envs.mdp.physics_events` when the active backend is Newton.
