+++ source/isaaclab_ovphysx/changelog.d/zhengyuz-decouple-physics-events.rst
@@ -1,0 +1,6 @@
+Added
+^^^^^
+
+* Added OVPhysX backend implementations of the MDP physics event-randomization terms under
+  ``isaaclab_ovphysx.envs.mdp``. OVPhysX reuses the PhysX implementations except for
+  ``randomize_rigid_body_material``, which is a no-op until the OVPhysX material binding is available.
