+++ source/isaaclab_physx/changelog.d/zhengyuz-decouple-physics-events.rst
@@ -1,0 +1,6 @@
+Added
+^^^^^
+
+* Added PhysX backend implementations of the MDP physics event-randomization terms under
+  ``isaaclab_physx.envs.mdp``. These are resolved at runtime by the converters in
+  :mod:`isaaclab.envs.mdp.physics_events` when the active backend is PhysX (and reused by OVPhysX).
